Refactoring of ExtendedRevisionCallGraph (WIP)
* Converts ExtendedRevisionCallGraph to an abstract class with generic
type A that is used as the type for CHA.
* Creates a new class ExtendedRevisionJavaCallGraph that implements
ExtendedRevisionCallGraph
* Creates separate classes fro Node, Type, Scope, Graph
* Graph can be reused in all supported languages
* Scope, and Type are currently used only by Java, thus we named them
JavaScope and JavaType respectively.
* Create a new abstract class for Node (TODO)
* Updates all tests
* Fix test testEmptyCallGraph of OPALPluginTest (TODO)
